+++ spec/grover/processor_spec.rb
@@ -8,13 +8,14 @@ describe Grover::Processor do
   describe '#convert' do
     subject(:convert) { processor.convert method, url_or_html, options }
 
-    let(:url_or_html) { 'http://google.com' }
+    let(:url_or_html) { 'http://localhost:4567' }
     let(:options) { {} }
     let(:date) do
       # New version of Chromium (v93) that comes with v10.2.0 of puppeteer uses a different date format
       date_format = puppeteer_version_on_or_after?('10.2.0') ? '%-m/%-d/%y, %-l:%M %p' : '%-m/%-d/%Y'
       Time.now.strftime date_format
     end
+    let(:protocol) { puppeteer_version_on_or_after?('21') ? 'https' : 'http' }
 
     context 'when converting to PDF' do
       let(:method) { :pdf }
@@ -27,7 +28,7 @@ describe Grover::Processor do
       let(:basic_header_footer_options) do
         {
           'displayHeaderFooter' => true,
-          'displayUrl' => 'http://www.example.net/foo/bar',
+          'displayUrl' => "#{protocol}://www.example.net/foo/bar",
           'margin' => {
             'top' => '1in',
             'bottom' => '1in'
@@ -40,14 +41,11 @@ describe Grover::Processor do
       end
 
       context 'when passing through a valid URL' do
-        # we need to add the language for test stability
-        # if not added explicitly, google can respond with a different locale
-        # based on IP address geo-lookup, timezone, etc.
-        let(:url_or_html) { 'https://www.google.com/?gl=us' }
+        let(:url_or_html) { 'http://localhost:4567' }
 
         it { is_expected.to start_with "%PDF-1.4\n" }
         it { expect(pdf_reader.page_count).to eq 1 }
-        it { expect(pdf_text_content).to include "I'm Feeling Lucky" }
+        it { expect(pdf_text_content).to include "I'm Feeling Grovery" }
       end
 
       context 'when passing through an invalid URL' do
@@ -60,6 +58,35 @@ describe Grover::Processor do
         end
       end
 
+      context 'when passing through a valid file URL' do
+        let(:url_or_html) { "file://#{fixture_path('test.html')}" }
+        let(:options) { { 'allowFileUri' => true } }
+
+        it { is_expected.to start_with "%PDF-1.4\n" }
+        it { expect(pdf_reader.page_count).to eq 1 }
+        it { expect(pdf_text_content).to include 'Hello World!' }
+      end
+
+      context 'when passing through an invalid file URL' do
+        let(:url_or_html) { 'file:///fake/invalid.html' }
+        let(:options) { { 'allowFileUri' => true } }
+
+        it 'raises a JavaScript error that the file could not be found' do
+          expect do
+            convert
+          end.to raise_error Grover::JavaScript::Error, %r{net::ERR_FILE_NOT_FOUND at file:///fake/invalid.html}
+        end
+      end
+
+      context 'when passing a file URL and allow_file_uri is disabled' do
+        let(:url_or_html) { "file://#{fixture_path('test.html')}" }
+
+        it { is_expected.to start_with "%PDF-1.4\n" }
+        it { expect(pdf_reader.page_count).to eq 1 }
+        it { expect(pdf_text_content).not_to include 'Hello World!' }
+        it { expect(pdf_text_content).to include "file://#{fixture_path('test.html')}" }
+      end
+
       context 'when passing through an empty string' do
         let(:url_or_html) { '' }
 
@@ -68,11 +95,21 @@ describe Grover::Processor do
         it { expect(pdf_text_content).to eq '' }
       end
 
-      context 'when puppeteer package is not installed' do
+      context 'when puppeteer / puppeteer-core package is not installed' do
         # Temporarily move the node puppeteer folder
-        before { FileUtils.move 'node_modules/puppeteer', 'node_modules/puppeteer_temp' }
+        before do
+          FileUtils.move 'node_modules/puppeteer', 'node_modules/puppeteer_temp'
+          if puppeteer_version_on_or_after? '18.0.0'
+            FileUtils.move 'node_modules/puppeteer-core', 'node_modules/puppeteer-core_temp'
+          end
+        end
 
-        after { FileUtils.move 'node_modules/puppeteer_temp', 'node_modules/puppeteer' }
+        after do
+          FileUtils.move 'node_modules/puppeteer_temp', 'node_modules/puppeteer'
+          if puppeteer_version_on_or_after? '18.0.0'
+            FileUtils.move 'node_modules/puppeteer-core_temp', 'node_modules/puppeteer-core'
+          end
+        end
 
         it 'raises a DependencyError' do
           expect { convert }.to raise_error Grover::DependencyError, Grover::Utils.squish(<<~ERROR)
@@ -84,7 +121,7 @@ describe Grover::Processor do
         context 'when puppeteer package is not in package.json' do
           before do
             FileUtils.copy 'package.json', 'package.json.tmp'
-            File.write('package.json', File.open('package.json') { |f| f.read.gsub(/"puppeteer"/, '"puppeteer-tmp"') })
+            File.write('package.json', File.open('package.json') { |f| f.read.gsub('"puppeteer"', '"puppeteer-tmp"') })
           end
 
           after { FileUtils.move 'package.json.tmp', 'package.json' }
@@ -98,6 +135,30 @@ describe Grover::Processor do
         end
       end
 
+      context 'when passing environment variables to the Node process through configuration' do
+        before { allow(Grover.configuration).to receive(:node_env_vars).and_return 'FOO' => 'bar' }
+
+        it 'passes the environment variables to the Node process' do
+          expect(Open3).to receive(:popen3).with({ 'FOO' => 'bar' }, any_args).and_call_original
+
+          convert
+        end
+      end
+
+      if puppeteer_version_on_or_after? '18.0.0'
+        context 'when only the puppeteer-core package is installed', :remote_browser do
+          before { FileUtils.move 'node_modules/puppeteer', 'node_modules/puppeteer_temp' }
+          after { FileUtils.move 'node_modules/puppeteer_temp', 'node_modules/puppeteer' }
+
+          let(:options) { { 'browserWsEndpoint' => browser_ws_endpoint } }
+          let(:browser_ws_endpoint) { 'ws://localhost:3000/' }
+          let(:url_or_html) { '<html><body style="background-color: blue">puppeteer-core works!</body></html>' }
+
+          it { expect { convert }.not_to raise_error }
+          it { expect(pdf_text_content).to eq 'puppeteer-core works!' }
+        end
+      end
+
       context 'when stubbing the call to the Node processor' do
         let(:stdin) { instance_double IO }
         let(:stdout) { instance_double IO }
@@ -107,7 +168,8 @@ describe Grover::Processor do
         before do
           allow(Open3).to(
             receive(:popen3).
-              with('node', File.expand_path(File.join(__dir__, '../../lib/grover/js/processor.cjs')), chdir: Dir.pwd).
+              with({}, 'node', File.expand_path(File.join(__dir__,
+                                                          '../../lib/grover/js/processor.cjs')), chdir: Dir.pwd).
               and_return([stdin, stdout, stderr, wait_thr])
           )
 
@@ -133,6 +195,8 @@ describe Grover::Processor do
 
         context 'when first call to gets on stdout succeeds but second returns nil' do
           before do
+            allow(stdin).to receive(:puts).with(any_args)
+
             allow(stdout).to receive(:gets).and_return '["ok"]', nil
             allow(stdin).to receive(:puts).with '["pdf","http://google.com",{}]'
             allow(stderr).to receive(:read).and_return 'The reason the worker failed'
@@ -149,7 +213,7 @@ describe Grover::Processor do
         context 'when first call to gets on stdout succeeds but second returns an error' do
           before do
             allow(stdout).to receive(:gets).and_return '["ok"]', '["err","Some unknown thing happened"]'
-            allow(stdin).to receive(:puts).with '["pdf","http://google.com",{}]'
+            allow(stdin).to receive(:puts).with '["pdf","http://localhost:4567",{}]'
             allow(stderr).to receive(:read).and_return 'The reason the worker failed'
           end
 
@@ -169,7 +233,7 @@ describe Grover::Processor do
         context 'when the worker returns an invalid response' do
           before do
             allow(stdout).to receive(:gets).and_return '["ok"]', '["ok",invalid_response]'
-            allow(stdin).to receive(:puts).with '["pdf","http://google.com",{}]'
+            allow(stdin).to receive(:puts).with '["pdf","http://localhost:4567",{}]'
           end
 
           it 'raises an Error' do
@@ -191,8 +255,11 @@ describe Grover::Processor do
 
         context 'when options includes A4 page format' do
           let(:options) { { format: 'A4' } }
+          let(:media_box) do
+            puppeteer_version_on_or_after?('21') ? [0, 0, 595.91998, 842.88] : [0, 0, 594.95996, 841.91998]
+          end
 
-          it { expect(pdf_reader.pages.first.attributes).to include(MediaBox: [0, 0, 594.95996, 841.91998]) }
+          it { expect(pdf_reader.pages.first.attributes).to include(MediaBox: media_box) }
         end
 
         context 'when options includes Letter page format' do
@@ -244,7 +311,7 @@ describe Grover::Processor do
         context 'when options include header and footer enabled' do
           let(:options) { basic_header_footer_options.merge('headerTemplate' => "#{large_text}#{default_header}") }
 
-          it { expect(pdf_text_content).to eq "#{date} Paaage Hey there http://www.example.net/foo/bar 1/1" }
+          it { expect(pdf_text_content).to eq "#{date} Paaage Hey there #{protocol}://www.example.net/foo/bar 1/1" }
         end
 
         context 'when options override header template' do
@@ -252,7 +319,7 @@ describe Grover::Processor do
             basic_header_footer_options.merge('headerTemplate' => "#{large_text}<div class='text'>Excellente</div>")
           end
 
-          it { expect(pdf_text_content).to eq 'Excellente Hey there http://www.example.net/foo/bar 1/1' }
+          it { expect(pdf_text_content).to eq "Excellente Hey there #{protocol}://www.example.net/foo/bar 1/1" }
         end
 
         context 'when header template includes the display url marker' do
@@ -264,7 +331,7 @@ describe Grover::Processor do
 
           it do
             expect(pdf_text_content).to(
-              eq('abchttp://www.example.net/foo/bardef Hey there http://www.example.net/foo/bar 1/1')
+              eq("abc#{protocol}://www.example.net/foo/bardef Hey there #{protocol}://www.example.net/foo/bar 1/1")
             )
           end
         end
@@ -273,7 +340,9 @@ describe Grover::Processor do
           let(:options) { basic_header_footer_options.merge('footerTemplate' => footer_template) }
           let(:footer_template) { "#{large_text}<div class='text'>great <span class='url'></span> page</div>" }
 
-          it { expect(pdf_text_content).to eq "#{date} Paaage Hey there great http://www.example.net/foo/bar page" }
+          it do
+            expect(pdf_text_content).to eq "#{date} Paaage Hey there great #{protocol}://www.example.net/foo/bar page"
+          end
 
           context 'when template contains quotes' do
             let(:footer_template) { %(<div class='text'>Footer with "quotes" in it</div>) }
@@ -284,9 +353,10 @@ describe Grover::Processor do
 
         context 'when displayUrl option is not provided' do
           let(:options) { basic_header_footer_options.tap { |hash| hash.delete('displayUrl') } }
+          let(:protocol) { puppeteer_version_on_or_after?('22') ? 'https' : 'http' }
 
           it 'uses the default `example.com` for the footer URL' do
-            expect(pdf_text_content).to eq "#{date} Paaage Hey there http://example.com/ 1/1"
+            expect(pdf_text_content).to eq "#{date} Paaage Hey there #{protocol}://example.com/ 1/1"
           end
         end
 
@@ -343,19 +413,32 @@ describe Grover::Processor do
 
         context 'when options include executablePath' do
           let(:options) { { 'executablePath' => '/totes/invalid/path' } }
+          let(:error_message) do
+            if puppeteer_version_on_or_after? '22.6.3'
+              'Browser was not found at the configured executablePath (/totes/invalid/path)'
+            else
+              %r{Failed to launch (chrome|the browser process)! spawn /totes/invalid/path}
+            end
+          end
 
           it do
             expect do
               convert
-            end.to raise_error Grover::JavaScript::Error,
-                               %r{Failed to launch (chrome|the browser process)! spawn /totes/invalid/path}
+            end.to raise_error Grover::JavaScript::Error, error_message
           end
         end
 
         context 'when requesting a URI requiring basic authentication' do
-          let(:url_or_html) { 'https://jigsaw.w3.org/HTTP/Basic/' }
+          let(:url_or_html) { 'http://localhost:4567/auth' }
 
-          it { expect(pdf_text_content).to eq 'Unauthorized access You are denied access to this resource.' }
+          if puppeteer_version_on_or_after? '22'
+            it do
+              expect { convert }.to raise_error Grover::JavaScript::Error,
+                                                'net::ERR_INVALID_AUTH_CREDENTIALS at http://localhost:4567/auth'
+            end
+          else
+            it { expect(pdf_text_content).to eq 'Unauthorized access You are denied access to this resource.' }
+          end
 
           context 'when passing through `username` and `password` options' do
             let(:options) { { username: 'guest', password: 'guest' } }
@@ -365,17 +448,17 @@ describe Grover::Processor do
         end
 
         context 'when passing through cookies option' do
-          let(:url_or_html) { 'https://cookierenderer-production.up.railway.app/' }
+          let(:url_or_html) { 'http://localhost:4567/cookie_renderer' }
           let(:options) do
             {
               'cookies' => [
                 {
                   'name' => 'grover-test',
                   'value' => 'nom nom nom',
-                  'domain' => 'cookierenderer-production.up.railway.app'
+                  'domain' => 'localhost:4567'
                 },
                 { 'name' => 'other-domain', 'value' => 'should not display', 'domain' => 'example.com' },
-                { 'name' => 'escaped', 'value' => '%26%3D%3D', 'domain' => 'cookierenderer-production.up.railway.app' }
+                { 'name' => 'escaped', 'value' => '%26%3D%3D', 'domain' => 'localhost:4567' }
               ]
             }
           end
@@ -386,25 +469,25 @@ describe Grover::Processor do
         end
 
         context 'when passing through extra HTTP headers' do
-          let(:url_or_html) { 'http://cookierenderer-production.up.railway.app/?type=headers' }
+          let(:url_or_html) { 'http://localhost:4567/headers' }
           let(:options) { { 'extraHTTPHeaders' => { 'grover-test' => 'yes it is' } } }
 
-          it { expect(pdf_text_content).to match(/Request contained (15|16) headers/) }
-          it { expect(pdf_text_content).to include '1. host cookierenderer-production.up.railway.app' }
-          it { expect(pdf_text_content).to include '4. grover-test yes it is' }
+          it { expect(pdf_text_content).to match(/Request contained \d+ headers/) }
+          it { expect(pdf_text_content).to include '1. host localhost:4567' }
+          it { expect(pdf_text_content).to match(/\d{1,2}\. grover-test yes it is/) }
         end
 
         context 'when overloading the user agent' do
-          let(:url_or_html) { 'http://cookierenderer-production.up.railway.app/?type=headers' }
+          let(:url_or_html) { 'http://localhost:4567/headers' }
           let(:options) { { 'userAgent' => 'Grover user agent' } }
 
-          it { expect(pdf_text_content).to match(/Request contained (14|15) headers/) }
-          it { expect(pdf_text_content).to include '1. host cookierenderer-production.up.railway.app' }
-          it { expect(pdf_text_content).to include 'user-agent Grover user agent' }
+          it { expect(pdf_text_content).to match(/Request contained \d+ headers/) }
+          it { expect(pdf_text_content).to include '1. host localhost:4567' }
+          it { expect(pdf_text_content).to match(/[45]\. user-agent Grover user agent/) }
         end
       end
 
-      context 'when passing through launch params to remote browser', remote_browser: true do
+      context 'when passing through launch params to remote browser', :remote_browser do
         let(:options) { { 'browserWsEndpoint' => browser_ws_endpoint } }
         let(:browser_ws_endpoint) { 'ws://localhost:3000/' }
         let(:url_or_html) do
@@ -427,6 +510,18 @@ describe Grover::Processor do
           let(:browser_ws_endpoint) { 'ws://localhost:3000/?--disable-speech-api' }
 
           it { expect(pdf_text_content).to eq 'Speech recognition is not supported' }
+        end
+      end
+
+      unless linux_system? # It looks like the way a connection failure is handled is different on Linux systems
+        context 'when passing through WS launch params without a remote browser' do
+          let(:options) { { 'browserWsEndpoint' => 'ws://localhost:3000/' } }
+          let(:url_or_html) { '<html><body>WS endpoint doesnt exist</body></html>' }
+
+          it 'raises a WsConnectFailedError' do
+            expect { convert }.to raise_error Grover::JavaScript::WsConnectFailedError,
+                                              'Failed to connect to browser WS endpoint: ws://localhost:3000/'
+          end
         end
       end
 
@@ -494,7 +589,7 @@ describe Grover::Processor do
         let(:options) { basic_header_footer_options.merge('executeScript' => script) }
         let(:script) { 'document.getElementsByTagName("body")[0].innerText = "Some evaluated content"' }
 
-        it { expect(pdf_text_content).to eq "#{date} Some evaluated content http://www.example.net/foo/bar 1/1" }
+        it { expect(pdf_text_content).to eq "#{date} Some evaluated content #{protocol}://www.example.net/foo/bar 1/1" }
       end
 
       context 'when waitForSelector option is specified' do
@@ -513,7 +608,7 @@ describe Grover::Processor do
         end
         let(:options) { basic_header_footer_options.merge('waitForSelector' => 'h1') }
 
-        it { expect(pdf_text_content).to eq "#{date} Hey there http://www.example.net/foo/bar 1/1" }
+        it { expect(pdf_text_content).to eq "#{date} Hey there #{protocol}://www.example.net/foo/bar 1/1" }
       end
 
       context 'when waitForFunction option is specified' do
@@ -539,7 +634,7 @@ describe Grover::Processor do
           )
         end
 
-        it { expect(pdf_text_content).to eq "#{date} Hey there http://www.example.net/foo/bar 1/1" }
+        it { expect(pdf_text_content).to eq "#{date} Hey there #{protocol}://www.example.net/foo/bar 1/1" }
       end
 
       context 'when waitForFunction option is specified with options' do
@@ -570,7 +665,7 @@ describe Grover::Processor do
           )
         end
 
-        it { expect(pdf_text_content).to eq "#{date} Hello, world! http://www.example.net/foo/bar 1/1" }
+        it { expect(pdf_text_content).to eq "#{date} Hello, world! #{protocol}://www.example.net/foo/bar 1/1" }
 
         context 'when waiting for function takes too long' do
           let(:wait_function_timeout) { 100 }
@@ -593,6 +688,7 @@ describe Grover::Processor do
           let(:url_or_html) do
             <<-HTML
               <html>
+                <head><link rel="icon" href="data:;base64,iVBORw0KGgo="></head>
                 <body>
                   <img src="http://foo.bar/baz.img" />
                 </body>
@@ -603,7 +699,8 @@ describe Grover::Processor do
           it do
             expect do
               convert
-            end.to raise_error Grover::JavaScript::RequestFailedError, 'net::ERR_NAME_NOT_RESOLVED at http://foo.bar/baz.img'
+            end.to raise_error Grover::JavaScript::RequestFailedError,
+                               "net::ERR_NAME_NOT_RESOLVED at #{protocol}://foo.bar/baz.img"
           end
         end
 
@@ -611,17 +708,25 @@ describe Grover::Processor do
           let(:url_or_html) do
             <<-HTML
               <html>
+                <head><link rel="icon" href="data:;base64,iVBORw0KGgo="></head>
                 <body>
                   <img src="https://google.com/404.jpg" />
                 </body>
               </html>
             HTML
           end
+          let(:error_message) do
+            if puppeteer_version_on_or_after? '22.6.0'
+              'net::ERR_BLOCKED_BY_ORB at https://google.com/404.jpg'
+            else
+              '404 https://google.com/404.jpg'
+            end
+          end
 
           it do
             expect do
               convert
-            end.to raise_error Grover::JavaScript::RequestFailedError, '404 https://google.com/404.jpg'
+            end.to raise_error Grover::JavaScript::RequestFailedError, error_message
           end
         end
 
@@ -629,6 +734,7 @@ describe Grover::Processor do
           let(:url_or_html) do
             <<-HTML
               <html>
+                <head><link rel="icon" href="data:;base64,iVBORw0KGgo="></head>
                 <body>
                   Hey there
                   <img src="https://httpstat.us/304" />
@@ -641,15 +747,16 @@ describe Grover::Processor do
         end
 
         context 'when assets have redirects PDFs are generated successfully' do
-          it { expect(pdf_text_content).to match "#{date} Google" }
+          it { expect(pdf_text_content).to match "#{date} I'm Feeling Grovery" }
         end
 
         context 'with images' do
           let(:url_or_html) do
             <<-HTML
               <html>
+                <head><link rel='icon' href='data:;base64,iVBORw0KGgo='></head>
                 <body>
-                  <img src="https://placekitten.com/200/200" />
+                  <img src="http://localhost:4567/cat.png" />
                 </body>
               </html>
             HTML
@@ -685,15 +792,17 @@ describe Grover::Processor do
           )
         end
 
-        it { expect(pdf_text_content).to eq "#{date} http://www.example.net/foo/bar 1/1" }
+        it { expect(pdf_text_content).to eq "#{date} #{protocol}://www.example.net/foo/bar 1/1" }
       end
 
       # Only test `waitForTimeout` if the Puppeteer supports it
-      if puppeteer_version_on_or_after? '5.3.0'
+      if puppeteer_version_on_or_after?('5.3.0') && puppeteer_version_on_or_before?('21.11.0')
         context 'when waitForTimeout option is specified' do
           let(:url_or_html) do
             <<-HTML
               <html>
+                <head><link rel='icon' href='data:;base64,iVBORw0KGgo='></head>
+              #{'  '}
                 <body>
                   <p id="loading">Loading</p>
                   <p id="content" style="display: none">Loaded</p>
@@ -809,11 +918,14 @@ describe Grover::Processor do
             let(:timeout) { 1 }
 
             if puppeteer_version_on_or_after? '10.4.0'
-              it 'will timeout when trying to convert to PDF' do
-                expect { convert }.to raise_error(
-                  Grover::JavaScript::TimeoutError,
-                  'waiting for Page.printToPDF failed: timeout 1ms exceeded'
-                )
+              it 'times out when trying to convert to PDF' do
+                error_message = if puppeteer_version_on_or_after?('21')
+                                  'Timed out after waiting 1ms'
+                                else
+                                  'waiting for Page.printToPDF failed: timeout 1ms exceeded'
+                                end
+
+                expect { convert }.to raise_error Grover::JavaScript::TimeoutError, error_message
               end
             else
               it { is_expected.to start_with "%PDF-1.4\n" }
@@ -830,11 +942,13 @@ describe Grover::Processor do
           let(:convert_timeout) { 1 }
 
           if puppeteer_version_on_or_after? '10.4.0'
-            it 'will raise an error when trying to convert to PDF' do
-              expect { convert }.to raise_error(
-                Grover::JavaScript::TimeoutError,
-                'waiting for Page.printToPDF failed: timeout 1ms exceeded'
-              )
+            it 'raises an error when trying to convert to PDF' do
+              error_message = if puppeteer_version_on_or_after?('21')
+                                'Timed out after waiting 1ms'
+                              else
+                                'waiting for Page.printToPDF failed: timeout 1ms exceeded'
+                              end
+              expect { convert }.to raise_error Grover::JavaScript::TimeoutError, error_message
             end
           else
             it { is_expected.to start_with "%PDF-1.4\n" }
@@ -844,11 +958,13 @@ describe Grover::Processor do
             let(:timeout) { 10_000 }
 
             if puppeteer_version_on_or_after? '10.4.0'
-              it 'will use the convert timeout over the timeout option' do
-                expect { convert }.to raise_error(
-                  Grover::JavaScript::TimeoutError,
-                  'waiting for Page.printToPDF failed: timeout 1ms exceeded'
-                )
+              it 'uses the convert timeout over the timeout option' do
+                error_message = if puppeteer_version_on_or_after?('21')
+                                  'Timed out after waiting 1ms'
+                                else
+                                  'waiting for Page.printToPDF failed: timeout 1ms exceeded'
+                                end
+                expect { convert }.to raise_error Grover::JavaScript::TimeoutError, error_message
               end
             else
               it { is_expected.to start_with "%PDF-1.4\n" }
@@ -876,7 +992,7 @@ describe Grover::Processor do
       let(:image) { MiniMagick::Image.read convert }
 
       context 'when passing through a valid URL' do
-        let(:url_or_html) { 'https://media.gettyimages.com/photos/tabby-cat-selfie-picture-id1151094724?s=2048x2048' }
+        let(:url_or_html) { 'http://localhost:4567/cat' }
 
         # default screenshot is PNG 800w x 600h
         it { expect(convert.unpack('C*')).to start_with "\x89PNG\r\n\x1A\n".unpack('C*') }
@@ -886,9 +1002,10 @@ describe Grover::Processor do
         # don't really want to rely on pixel testing the website screenshot
         # so we'll check it's mean colour is roughly what we expect
         it do
+          pixel_mean = puppeteer_version_on_or_after?('22.9.0') ? 140.925 : 161.497
+
           expect(image.data.dig('imageStatistics', MiniMagick.imagemagick7? ? 'Overall' : 'all', 'mean').to_f).
-            to be_within(1).of(97.7473).  # ImageMagick 6.9.3-1 (version used by Travis CI)
-            or be_within(1).of(161.497)   # ImageMagick 6.9.10-84
+            to be_within(10).of(pixel_mean)
         end
       end
 
@@ -901,7 +1018,7 @@ describe Grover::Processor do
         it { expect(mean_colour_statistics(image)).to eq %w[0 0 255] }
       end
 
-      context 'when using remote browser option with HTML', remote_browser: true do
+      context 'when using remote browser option with HTML', :remote_browser do
         let(:options) { { browserWsEndpoint: 'ws://localhost:3000' } }
         let(:url_or_html) { '<html><body style="background-color: blue"></body></html>' }